Explicitly downgrade errors to warnings
Previously, errors would simply be added to Webpack's list of
warnings. With this commit, files that include errors
actually no longer get marked as such (when `emitErrors` is
false). This results in stylelint errors actually being
added to the list of warnings, rather than to the list
of errors, removing the need to special case them when compiling
the compilation stats.

Since the `severity` of these errors is still set to `error`,
this will still be indicated to the user - it just will no longer
abort compilation.